feat(form creating words): create form for creating words
