+++ src/modules/create-words/components/table-words/TableWords.tsx
@@ -1,0 +1,138 @@
+import React, { useState } from 'react';
+import { TableProps } from 'antd';
+import { ColumnType } from 'antd/es/table';
+import type { FilterConfirmProps } from 'antd/es/table/interface';
+import { Button } from '@/ui-components/Button';
+import { Table } from '@/ui-components/Table';
+import { Input } from '@/ui-components/Input';
+import { Space } from '@/ui-components/Space';
+import { Icon } from '@/ui-components/Icon';
+import { DataWords } from '../main/CreateWords';
+import styles from './tableWords.module.css';
+
+type TableRowSelection<T extends object = object> = TableProps<T>['rowSelection'];
+
+export const TableWords = ({
+	tableWords,
+	setTableWords,
+}: {
+	tableWords: DataWords[];
+	setTableWords: React.Dispatch<React.SetStateAction<DataWords[]>>;
+}) => {
+	const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([]);
+	const [, setSearchText] = useState('');
+	const [, setSearchedColumn] = useState<keyof DataWords | ''>('');
+
+	const onSelectChange = (newSelectedRowKeys: React.Key[]) => {
+		setSelectedRowKeys(newSelectedRowKeys);
+	};
+
+	const deleteRows = () => {
+		const newRows = tableWords.filter((row) => !selectedRowKeys.includes(row.key));
+		setTableWords(newRows);
+	};
+
+	const handleSearch = (
+		selectedKeys: string[],
+		confirm: (param?: FilterConfirmProps) => void,
+		dataIndex: keyof DataWords,
+	) => {
+		confirm();
+		setSearchText(selectedKeys[0]);
+		setSearchedColumn(dataIndex);
+	};
+
+	const handleReset = (clearFilters?: () => void) => {
+		clearFilters?.();
+		setSearchText('');
+	};
+
+	const getColumnSearchProps = (dataIndex: keyof DataWords): ColumnType<DataWords> => ({
+		filterDropdown: ({ setSelectedKeys, selectedKeys, confirm, clearFilters }) => (
+			<div style={{ padding: 8 }}>
+				<Input
+					placeholder={`Search ${dataIndex}`}
+					value={selectedKeys[0]}
+					onChange={(e) => setSelectedKeys(e.target.value ? [e.target.value] : [])}
+					onPressEnter={() => handleSearch(selectedKeys as string[], confirm, dataIndex)}
+					style={{ marginBottom: 8, display: 'block' }}
+				/>
+				<Space>
+					<Button
+						type="primary"
+						onClick={() => handleSearch(selectedKeys as string[], confirm, dataIndex)}
+						icon={<Icon icon="search" />}
+						size="small"
+						style={{ width: 90 }}
+					>
+						Search
+					</Button>
+					<Button onClick={() => handleReset(clearFilters)} size="small" style={{ width: 90 }}>
+						Reset
+					</Button>
+				</Space>
+			</div>
+		),
+		filterIcon: (filtered: boolean) => (
+			<Icon icon="search" size="xs" variant={filtered ? 'primary' : 'secondary'} />
+		),
+		onFilter: (value, record) =>
+			record[dataIndex]
+				?.toString()
+				.toLowerCase()
+				.includes((value as string).toLowerCase()),
+	});
+
+	const columns = [
+		Table.EXPAND_COLUMN,
+		{
+			title: 'Name',
+			dataIndex: 'name',
+			key: 'name',
+			sorter: (a: DataWords, b: DataWords) => a.name.localeCompare(b.name),
+			...getColumnSearchProps('name'),
+		},
+		{
+			title: 'Transcription',
+			dataIndex: 'transcription',
+			key: 'transcription',
+		},
+		{
+			title: 'Translate',
+			dataIndex: 'translate',
+			key: 'translate',
+		},
+		{
+			title: 'Type',
+			dataIndex: 'type',
+			key: 'type',
+		},
+		{
+			title: 'Variants',
+			dataIndex: 'variants',
+			key: 'variants',
+		},
+	];
+
+	const rowSelection: TableRowSelection<DataWords> = {
+		selectedRowKeys,
+		onChange: onSelectChange,
+	};
+	return (
+		<div className={styles.tableContainer}>
+			<div className={styles.deleteButtonContainer}>
+				<Button onClick={deleteRows}>Delete</Button>
+			</div>
+			<Table<DataWords>
+				className={styles.table}
+				size="small"
+				expandable={{
+					expandedRowRender: (record) => <p className={styles.description}>{record.useCase}</p>,
+				}}
+				dataSource={tableWords}
+				columns={columns}
+				rowSelection={rowSelection}
+			/>
+		</div>
+	);
+};
