+++ src/modules/create-words/components/table-words/tableWords.module.css
@@ -1,0 +1,26 @@
+.tableContainer {
+	margin-top: 30px;
+	width: 100%;
+	overflow-x: auto;
+	display: flex;
+	flex-direction: column;
+	gap: 20px;
+}
+
+.table {
+	width: 100%;
+}
+
+.table :global(.ant-table-thead > tr > th) {
+	background-color: #f0f2f5;
+	color: #000;
+}
+
+.table :global(.ant-table-expanded-row) {
+	background-color: #f0f2f5;
+	padding: 16px;
+}
+
+.description {
+	margin: 0;
+}
