+++ src/modules/create-words/types.ts
@@ -1,6 +1,6 @@
 import z from 'zod';
 import { FiltersDto } from '@/shared/api/generated/model';
-import { schema } from './zod-schemas/form.schema';
+import { dataWordSchema } from './zod-schemas/form.schema';
 
-export type FormValues = z.infer<typeof schema>;
+export type FormValues = z.infer<typeof dataWordSchema>;
 export type FiltersKeys = keyof FiltersDto;
